Cập nhật code    + Type ServiceType    + Function tạo order    + Cập nhật database Thêm    + OrderType    + OrderItemType

+++ app/GraphQL/Types/OptionType.php
@@ -40,7 +40,7 @@ class OptionType extends GraphQLType
                 'description' => 'The default of the option',
             ],
             'price' => [
-                'type' => Type::Double(),
+                'type' => Type::float(),
                 'description' => 'The index of the option',
             ],
             'extra' => [
